Avoid double counting travel time on custom itineraries
-the original itineraries were generated using a convention where a
generated event (such as a sleep or food event) would share locations
with the event coming immediately after it. this convention isn't
well suited for loading custom itineraries from data files, as it is
much cleaner to use the location of the last event as you read the
data file (instead of forcing the file reader to try to read the
next line and then go back and update the generated event)
-the issue here was that when a generated itinerary was loaded as a
custom itinerary, the travel time between the generated event and the revious event had already been added when the itinerary was
generated. however, when the itinerary was loaded from the file, the
generated event was now sharing a location with the previous event
and an additional travel time between the generated event and the
next event was added. in the end generated events had the full travel
time both before and after the event.
-this hack allows us to remove the initially added travel time between the generated event and the preivous event.